requirements.txt
websocket-client>=1.8.0
pandas>=2.2.3
numpy>=2.1.0
ta>=0.11.0
pymongo>=4.7.0
dnspython>=2.6.0
certifi>=2024.2.2
python-dotenv>=1.0.1
apscheduler>=3.10.4
fastapi>=0.115.0
uvicorn>=0.32.0
loguru>=0.7.2
backtrader>=1.9.78.123
matplotlib>=3.9.0
pytest>=8.3.0
httpx>=0.27.0
MetaTrader5>=5.0.0
tabulate>=0.9.0